moving json encoder/decoder to json_utils.py
